chore: depend on @gmod/bgzf-filehandle 6.3.2
Which annotates unzip() as Promise<Uint8Array> instead of leaving it to be
inferred as `any` through the inlined-wasm bundle, and exports the
unzipChunkSlice result as a named type. That retires the two `as Uint8Array`
assertions that stood in for the annotation, and lets ReadChunk be the
package's own ChunkSlice rather than a restatement of it that could drift.

Exempts @gmod/* from the install quarantine, as jbrowse-components already
does: a bump here is normally made minutes after the dependency's own release,
and `pnpm install --frozen-lockfile` in CI would otherwise fail for three days.

+++ src/indexFile.ts
@@ -89,9 +89,7 @@ export default abstract class IndexFile {
       signal: opts.signal,
       onProgress: opts.onProgress,
     })
-    // the assertion is for @gmod/bgzf-filehandle <= 6.3.1, which infers unzip()
-    // as `any`; it can go once this depends on a release that annotates it
-    const bytes = (await unzip(buf)) as Uint8Array
+    const bytes = await unzip(buf)
     return {
       bytes,
       dataView: new DataView(bytes.buffer, bytes.byteOffset, bytes.byteLength),
